FF: Code to make Routine Settings button wasn't 3.11 compatible

+++ psychopy/app/builder/builder.py
@@ -2273,7 +2273,7 @@ class RoutineCanvas(wx.ScrolledWindow, handlers.ThemeMixin):
             self.componentFromID[id] = component
         dc.SetId(id)
         # Get settings icon
-        sz = self.iconSize/3
+        sz = int(self.iconSize/3)
         thisIcon = icons.ComponentIcon(component, size=sz).bitmap
         # Some parameters
         lbl = _translate("Routine settings")
